EPMRPP-92671 || code review fix - 1
